+++ src/zone-inl.h
@@ -100,7 +100,7 @@ void* ZoneObject::operator new(size_t size, Zone* zone) {
 
 inline void* ZoneAllocationPolicy::New(size_t size) {
   ASSERT(zone_);
-  return zone_->New(size);
+  return zone_->New(static_cast<int>(size));
 }
